Merges reconcilers
